+++ src/Models/Wallet.php
@@ -4,7 +4,7 @@ namespace Bavix\Wallet\Models;
 
 use Bavix\Wallet\Interfaces\Customer;
 use Bavix\Wallet\Interfaces\WalletFloat;
-use Bavix\Wallet\Traits\CanBePaidFloat;
+use Bavix\Wallet\Traits\CanPayFloat;
 use Bavix\Wallet\Traits\HasGift;
 use Bavix\Wallet\WalletProxy;
 use Illuminate\Database\Eloquent\Model;
@@ -21,7 +21,7 @@ use Illuminate\Support\Str;
 class Wallet extends Model implements Customer, WalletFloat
 {
 
-    use CanBePaidFloat;
+    use CanPayFloat;
     use HasGift;
 
     /**